Update index.html
Updated file to include year: 2025 (the current year). But, I haven't added any class schedule, yet...; because, I'm still not exactly sure what the timetable will be...; there are still 3 x online 1 & 1/2 hour long applications to make...? Depending on if I pass or not...; then, I will we allowed to attend that class/or, not.